+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the `dom_query` crate will be documented in this file.
 
 ### Added
 - Introduced `Tree::html_root` and `Document::html_root` methods to get the root element (`<html>`) node of a document.
+- Implemented the `NodeRef::to_fragment` method to create a full copy of a node's contents as a `Document` fragment.
 
 ## [0.18.0] - 2025-04-26