+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the `dom_query` crate will be documented in this file.
 
-## [Unreleased]
+## [0.19.0] - 2025-05-20
 
 ### Added
 - Introduced `Tree::html_root` and `Document::html_root` methods to get the root element (`<html>`) node of a document.